refactor Debug logging to a module shared with Analysers
